backend/requirements.txt
# Core
Django==6.0
djangorestframework==3.16.1
psycopg==3.3.2
psycopg-binary==3.3.2

# Utilities
python-dotenv==1.2.1
dj-database-url==3.0.1

# Security
django-axes==6.5.1
django-ratelimit==4.1.0
django-cors-headers==4.3.1

# Development
pyright==1.1.407
